stb image moved to externals
