+++ externals/stb_image.h
